+++ src/Credfeto.DotNet.Repo.Tools.CleanUp.Tests/Services/ProjectXmlRewriterTests.PropertyGroup.cs
@@ -10,8 +10,7 @@ public sealed partial class ProjectXmlRewriterTests
     [Fact]
     public Task ReOrderPropertieShouldNotChangeAnythingWhenCommentsDetectedAsync()
     {
-        const string originalXml =
-            @"<Project Sdk=""Microsoft.NET.Sdk"">
+        const string originalXml = @"<Project Sdk=""Microsoft.NET.Sdk"">
   <PropertyGroup>
     <DisableImplicitNuGetFallbackFolder>true</DisableImplicitNuGetFallbackFolder>
     <EnablePackageValidation>true</EnablePackageValidation>
@@ -101,8 +100,7 @@ public sealed partial class ProjectXmlRewriterTests
   </ItemGroup>
 </Project>";
 
-        const string expectedXml =
-            @"<Project Sdk=""Microsoft.NET.Sdk"">
+        const string expectedXml = @"<Project Sdk=""Microsoft.NET.Sdk"">
   <PropertyGroup>
     <DisableImplicitNuGetFallbackFolder>true</DisableImplicitNuGetFallbackFolder>
     <EnablePackageValidation>true</EnablePackageValidation>
@@ -192,14 +190,13 @@ public sealed partial class ProjectXmlRewriterTests
   </ItemGroup>
 </Project>";
 
-        return this.DoReOorderPropertiesAsync(expectedXml: expectedXml, originalXml: originalXml);
+        return this.DoReOrderPropertiesAsync(expectedXml: expectedXml, originalXml: originalXml);
     }
 
     [Fact]
     public Task ReOrderPropertieShouldSortButNotCombineWhenAttributesDetectedAsync()
     {
-        const string originalXml =
-            @"<Project Sdk=""Microsoft.NET.Sdk"">
+        const string originalXml = @"<Project Sdk=""Microsoft.NET.Sdk"">
   <PropertyGroup Test=""True"">
     <DisableImplicitNuGetFallbackFolder>true</DisableImplicitNuGetFallbackFolder>
     <EnablePackageValidation>true</EnablePackageValidation>
@@ -286,8 +283,7 @@ public sealed partial class ProjectXmlRewriterTests
   </ItemGroup>
 </Project>";
 
-        const string expectedXml =
-            @"<Project Sdk=""Microsoft.NET.Sdk"">
+        const string expectedXml = @"<Project Sdk=""Microsoft.NET.Sdk"">
   <PropertyGroup Test=""True"">
     <DebuggerSupport>true</DebuggerSupport>
     <DisableImplicitNuGetFallbackFolder>true</DisableImplicitNuGetFallbackFolder>
@@ -374,14 +370,13 @@ public sealed partial class ProjectXmlRewriterTests
   </ItemGroup>
 </Project>";
 
-        return this.DoReOorderPropertiesAsync(expectedXml: expectedXml, originalXml: originalXml);
+        return this.DoReOrderPropertiesAsync(expectedXml: expectedXml, originalXml: originalXml);
     }
 
     [Fact]
     public Task ReOrderPropertieShouldChangeToAlphanumericWhenNoCommentsDetectedAsync()
     {
-        const string originalXml =
-            @"<Project Sdk=""Microsoft.NET.Sdk"">
+        const string originalXml = @"<Project Sdk=""Microsoft.NET.Sdk"">
   <PropertyGroup>
     <DisableImplicitNuGetFallbackFolder>true</DisableImplicitNuGetFallbackFolder>
     <EnablePackageValidation>true</EnablePackageValidation>
@@ -468,8 +463,7 @@ public sealed partial class ProjectXmlRewriterTests
   </ItemGroup>
 </Project>";
 
-        const string expectedXml =
-            @"<Project Sdk=""Microsoft.NET.Sdk"">
+        const string expectedXml = @"<Project Sdk=""Microsoft.NET.Sdk"">
   <PropertyGroup>
     <AnalysisLevel>latest</AnalysisLevel>
     <AnalysisMode>AllEnabledByDefault</AnalysisMode>
@@ -552,6 +546,6 @@ public sealed partial class ProjectXmlRewriterTests
   </ItemGroup>
 </Project>";
 
-        return this.DoReOorderPropertiesAsync(expectedXml: expectedXml, originalXml: originalXml);
+        return this.DoReOrderPropertiesAsync(expectedXml: expectedXml, originalXml: originalXml);
     }
 }