+++ src/Credfeto.DotNet.Repo.Tools.CleanUp/Services/LoggingExtensions/BulkCodeCleanUpLoggingExtensions.cs
@@ -22,16 +22,8 @@ internal static partial class BulkCodeCleanUpLoggingExtensions
     [LoggerMessage(EventId = 5, Level = LogLevel.Information, Message = "[CACHE] Updating {packageId}")]
     public static partial void LogUpdatingCachedPackage(this ILogger<BulkCodeCleanUp> logger, string packageId);
 
-    [LoggerMessage(
-        EventId = 6,
-        Level = LogLevel.Information,
-        Message = "[CACHE] Update {packageId} updated {count} packages"
-    )]
-    public static partial void LogUpdatedCachedPackages(
-        this ILogger<BulkCodeCleanUp> logger,
-        string packageId,
-        int count
-    );
+    [LoggerMessage(EventId = 6, Level = LogLevel.Information, Message = "[CACHE] Update {packageId} updated {count} packages")]
+    public static partial void LogUpdatedCachedPackages(this ILogger<BulkCodeCleanUp> logger, string packageId, int count);
 
     [LoggerMessage(EventId = 7, Level = LogLevel.Information, Message = "[CACHE] Total package updates: {count}")]
     public static partial void LogUpdatedCachedPackagesTotal(this ILogger<BulkCodeCleanUp> logger, int count);
@@ -46,113 +38,41 @@ internal static partial class BulkCodeCleanUpLoggingExtensions
     public static partial void LogNoDotNetFilesFound(this ILogger<BulkCodeCleanUp> logger);
 
     [LoggerMessage(EventId = 11, Level = LogLevel.Information, Message = "Resetting {clonePath} to {branch}")]
-    private static partial void LogResettingToDefault(
-        this ILogger<BulkCodeCleanUp> logger,
-        string clonePath,
-        string branch
-    );
+    private static partial void LogResettingToDefault(this ILogger<BulkCodeCleanUp> logger, string clonePath, string branch);
 
     public static void LogResettingToDefault(this ILogger<BulkCodeCleanUp> logger, in RepoContext repoContext)
     {
         logger.LogResettingToDefault(clonePath: repoContext.ClonePath, branch: repoContext.DefaultBranch);
     }
 
-    [LoggerMessage(
-        EventId = 12,
-        Level = LogLevel.Information,
-        Message = "{clonePath}: Committing {packageId} ({version}) to {branch}"
-    )]
-    private static partial void LogCommittingPackageToBranch(
-        this ILogger<BulkCodeCleanUp> logger,
-        string clonePath,
-        string packageId,
-        string version,
-        string branch
-    );
+    [LoggerMessage(EventId = 12, Level = LogLevel.Information, Message = "{clonePath}: Committing {packageId} ({version}) to {branch}")]
+    private static partial void LogCommittingPackageToBranch(this ILogger<BulkCodeCleanUp> logger, string clonePath, string packageId, string version, string branch);
 
-    public static void LogCommittingToDefault(
-        this ILogger<BulkCodeCleanUp> logger,
-        in RepoContext repoContext,
-        string packageId,
-        string version
-    )
+    public static void LogCommittingToDefault(this ILogger<BulkCodeCleanUp> logger, in RepoContext repoContext, string packageId, string version)
     {
-        logger.LogCommittingPackageToBranch(
-            clonePath: repoContext.ClonePath,
-            packageId: packageId,
-            version: version,
-            branch: repoContext.DefaultBranch
-        );
+        logger.LogCommittingPackageToBranch(clonePath: repoContext.ClonePath, packageId: packageId, version: version, branch: repoContext.DefaultBranch);
     }
 
-    public static void LogCommittingToNamedBranch(
-        this ILogger<BulkCodeCleanUp> logger,
-        in RepoContext repoContext,
-        string branch,
-        string packageId,
-        string version
-    )
+    public static void LogCommittingToNamedBranch(this ILogger<BulkCodeCleanUp> logger, in RepoContext repoContext, string branch, string packageId, string version)
     {
-        logger.LogCommittingPackageToBranch(
-            clonePath: repoContext.ClonePath,
-            packageId: packageId,
-            version: version,
-            branch: branch
-        );
+        logger.LogCommittingPackageToBranch(clonePath: repoContext.ClonePath, packageId: packageId, version: version, branch: branch);
     }
 
-    [LoggerMessage(
-        EventId = 13,
-        Level = LogLevel.Information,
-        Message = "{clonePath}: Skipping commit of {packageId} {version} as branch {branch} already exists"
-    )]
-    private static partial void LogSkippingPackageCommit(
-        this ILogger<BulkCodeCleanUp> logger,
-        string clonePath,
-        string packageId,
-        string version,
-        string branch
-    );
+    [LoggerMessage(EventId = 13, Level = LogLevel.Information, Message = "{clonePath}: Skipping commit of {packageId} {version} as branch {branch} already exists")]
+    private static partial void LogSkippingPackageCommit(this ILogger<BulkCodeCleanUp> logger, string clonePath, string packageId, string version, string branch);
 
-    public static void LogSkippingPackageCommit(
-        this ILogger<BulkCodeCleanUp> logger,
-        in RepoContext repoContext,
-        string branch,
-        string packageId,
-        string version
-    )
+    public static void LogSkippingPackageCommit(this ILogger<BulkCodeCleanUp> logger, in RepoContext repoContext, string branch, string packageId, string version)
     {
-        logger.LogSkippingPackageCommit(
-            clonePath: repoContext.ClonePath,
-            packageId: packageId,
-            version: version,
-            branch: branch
-        );
+        logger.LogSkippingPackageCommit(clonePath: repoContext.ClonePath, packageId: packageId, version: version, branch: branch);
     }
 
     [LoggerMessage(EventId = 18, Level = LogLevel.Warning, Message = "{message}")]
-    public static partial void LogReleaseCreated(
-        this ILogger<BulkCodeCleanUp> logger,
-        string message,
-        Exception exception
-    );
+    public static partial void LogReleaseCreated(this ILogger<BulkCodeCleanUp> logger, string message, Exception exception);
 
-    [LoggerMessage(
-        EventId = 19,
-        Level = LogLevel.Warning,
-        Message = "SDK {sdkVersion} was requested, but not installed.  Currently installed SDKS: {installedSdks}"
-    )]
-    private static partial void LogMissingSdk(
-        this ILogger<BulkCodeCleanUp> logger,
-        string sdkVersion,
-        string installedSdks
-    );
+    [LoggerMessage(EventId = 19, Level = LogLevel.Warning, Message = "SDK {sdkVersion} was requested, but not installed.  Currently installed SDKS: {installedSdks}")]
+    private static partial void LogMissingSdk(this ILogger<BulkCodeCleanUp> logger, string sdkVersion, string installedSdks);
 
-    public static void LogMissingSdk(
-        this ILogger<BulkCodeCleanUp> logger,
-        Version sdkVersion,
-        IReadOnlyList<Version> installedSdks
-    )
+    public static void LogMissingSdk(this ILogger<BulkCodeCleanUp> logger, Version sdkVersion, IReadOnlyList<Version> installedSdks)
     {
         logger.LogMissingSdk(sdkVersion.ToString(), string.Join(separator: ", ", values: installedSdks));
     }
@@ -160,23 +80,18 @@ internal static partial class BulkCodeCleanUpLoggingExtensions
     [LoggerMessage(EventId = 20, Level = LogLevel.Information, Message = "Project Cleanup Starting: {filename}")]
     public static partial void StartingProjectCleaup(this ILogger<BulkCodeCleanUp> logger, string filename);
 
-    [LoggerMessage(
-        EventId = 21,
-        Level = LogLevel.Information,
-        Message = "Project Cleanup Completed: {filename} Changes: {changes}"
-    )]
-    public static partial void CompletingProjectCleaup(
-        this ILogger<BulkCodeCleanUp> logger,
-        string filename,
-        int changes
-    );
+    [LoggerMessage(EventId = 21, Level = LogLevel.Information, Message = "Project Cleanup Completed: {filename} Changes: {changes}")]
+    public static partial void CompletingProjectCleanup(this ILogger<BulkCodeCleanUp> logger, string filename, int changes);
 
-    [LoggerMessage(EventId = 22, Level = LogLevel.Information, Message = "Cleaning: {filename}")]
+    [LoggerMessage(EventId = 22, Level = LogLevel.Information, Message = "Project Cleanup Failed: {filename} Status: {message}")]
+    public static partial void FailedProjectCleanup(this ILogger<BulkCodeCleanUp> logger, string filename, string message, Exception exception);
+
+    [LoggerMessage(EventId = 23, Level = LogLevel.Information, Message = "Cleaning: {filename}")]
     public static partial void CleaningFile(this ILogger<BulkCodeCleanUp> logger, string filename);
 
-    [LoggerMessage(EventId = 23, Level = LogLevel.Information, Message = "Cleaning: {filename} (unchanged)")]
+    [LoggerMessage(EventId = 24, Level = LogLevel.Information, Message = "Cleaning: {filename} (unchanged)")]
     public static partial void CleaningFileUnchanged(this ILogger<BulkCodeCleanUp> logger, string filename);
 
-    [LoggerMessage(EventId = 24, Level = LogLevel.Information, Message = "Cleaning: {filename} (introduced changes)")]
+    [LoggerMessage(EventId = 25, Level = LogLevel.Information, Message = "Cleaning: {filename} (introduced changes)")]
     public static partial void CleaningFileDifferent(this ILogger<BulkCodeCleanUp> logger, string filename);
 }