+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ## [Unreleased]
 ### Added
 ### Fixed
+
+- Unhandled exception in xml rewriter
 ### Changed
 ### Removed
 ### Deployment Changes